requirements_ingest.txt
paho-mqtt
psycopg2-binary